cmd: update usage
